requirements (1).txt
streamlit==1.23.1
pandas==2.0.2
numpy==1.24.3
Pillow==9.5.0
Levenshtein==0.21.1
openpyxl==3.1.2
